add sling_conns.go
